requirements.txt
requests
pandas
sqlalchemy
psycopg2-binary
python-dotenv
boto3
matplotlib
